test-requirements.txt
# Install bounded pep8/pyflakes first, then let flake8 install
pep8==1.4.5
pyflakes==0.7.2
flake8==2.0
hacking>=0.5.3,<0.6

discover
fixtures>=0.3.12
python-subunit
testrepository>=0.0.17
testscenarios<0.5
testtools>=0.9.29
WebOb>=1.2.3,<1.3

# when we can require tox>= 1.4, this can go into tox.ini:
#  [testenv:cover]
#  deps = {[testenv]deps} coverage
coverage

# this is required for the docs build jobs
sphinx
oslo.sphinx
